requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
numpy
gradio
requests